moving atuin AFTER starship
On talapas (only?), doing atuin BEFORE starship caused starship to not
work at all. Moving atuin AFTER starship caused it to start working
again.

Seemed to only happen on Talapas (???), not laptop, others. Weird.

Might be entangled with my newfangled manual PROMPT_COMMAND as well
(the one with flock), as this change was made around the same time
as well. That said, my starship on Talapas was broken for months
up until today (because of a workshop back in summer where I had
disabled it). So who knows.